pyproject.toml
[project]
name = "ai-avatar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyttsx3",
    "gTTS",
    "langchain-groq>=1.1.2",
    "python-dotenv"
]
